pyproject.toml
# pyproject.toml
# =========================================
# Project: ccutils
# =========================================

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ccutils"
version = "0.1.0"
description = "Cookiecutter utilities for jcook3701's cookiecutter templates."
authors = [
  { name = "Jared Cook", email = "jcook3701+github@gmail.com" }
]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.11"
dependencies = [
  "cookiecutter",
  "GitPython",
  "typer",
  "types-requests",
]
keywords = [
  "python3",
  "cookiecutter",
  "templates",
  "utilities",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

[tool.setuptools.packages.find]
where = ["src"]
include = [
  "src",
  "tests",
]

[project.urls]
Homepage = "https://github.com/jcook3701/ccutils"
Repository = "https://github.com/jcook3701/ccutils"
Documentation = "https://github.com/jcook3701/ccutils/docs/jekyll"

[project.optional-dependencies]
dev = [
  "ruff>=0.14.0",
  "mypy==1.9",
  "pytest>=8.0",
  "yamllint>=1.0",
  "myst-parser>=0.18"
]
docs = [
  "sphinx>=8.0",
  "sphinx-rtd-theme>=3.0",
  "sphinx-markdown-builder>=0.6",
  "sphinx-autodoc-typehints>=3.5",
]

# ====================================
# ruff (linter, formatter, import sorter)
# ====================================
[tool.ruff]
target-version = "py311"
# Directories to lint and format
src = [
  "src",
  "tests",
]
# Exclude these paths
exclude = [
  ".venv",
  "build",
  "dist",
  "__pycache__",
  "docs",
  "*.egg-info",
  ".git",
  ".github",
]
# Maximum line length
line-length = 120

[tool.ruff.lint]
fixable = ["ALL"]

# Enable linting rules
select = [
  "W",  # PyCodeStyle warnings
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort (import sorting)
  "N",  # pep8-naming
  "UP", # pyupgrade
  "B",  # flake8-bugbear
  "A",  # flake8-builtins
  "PL", # pylint-style checks
]

# Ignore specific rules
ignore = [
  "E501",  # line too long (handled by formatter)
  "PLR2004",  # magic value comparison
  "B008",  # function call in default argument
]
# ====================================
# Import sorting (Python helper scripts)
# ====================================
[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["ccutils"]
# ====================================
# Formatter configuration
# Ruff also formats YAML, TOML, and Markdown
# ====================================
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false
# ====================================
# Mypy (type checker) Configuration
# ====================================
[tool.mypy]
python_version = 3.11
mypy_path = "src"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
# ====================================
# Pytest configuration
# ====================================
[tool.pytest.ini_options]
minversion = "7.0"
addopts = ["-v", "--maxfail=1", "--disable-warnings"]
testpaths = ["tests"]
pythonpath = ["src"]
# ====================================
# Optional entry point (lets you run with `python -m ccutils`)
# ====================================
[project.scripts]
ccutils = "ccutils.cli:app"
